requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
emoji==1.5.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.1
gspread==4.0.1
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauth2client==3.0.0
oauthlib==3.1.1
pandas
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyshorteners==1.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.1.0
python-dateutil==2.8.2
python-decouple==3.4
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.25
tabulate==0.8.9
termcolor==1.1.0
tornado==6.1
tzlocal==3.0
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.1
